Reduce memory usage to solve queryingthedocument
Score: 60.00  Status: Accepted

Reduced memory usage by allocating a single buffer to hold all words,
instead of allocating individual buffers for each word.  This works well
because although a perverse document could contain a word as long as the
maximum document length minus one (for the period ending the single
sentence), the total storage for all words in the document cannot exceed
the size of the document.